🚚 chore: Rename web-rtc-test folder to web-rtc and clean up
- Removed dummy data from web-rtc components
- Updated navigation paths to reflect the new folder structure
