+++ .env
@@ -1,6 +1,2 @@
 REACT_APP_SERVER_URL=http://localhost:8080
-
-export const getServerUrl = () => {
-    // return "https://192.168.36.125:3000";
-    return "https://rtcserver.nemooceanacademy.com:3000";
-}
+REACT_APP_RTC_SERVER_URL=https://rtcserver.nemooceanacademy.com:3000
