+++ .gitignore
@@ -13,7 +13,7 @@
 
 # misc
 .DS_Store
-.env.local
+.env.local 
 .env.development.local
 .env.test.local
 .env.production.local
@@ -23,4 +23,4 @@ yarn-debug.log*
 yarn-error.log*
 
 .idea
-src/components/web-rtc-test/secret.js
+.env
