+++ src/components/web-rtc/serverUrl.js
@@ -1,0 +1,3 @@
+export const getServerUrl = () => {
+    return process.env.REACT_APP_RTC_SERVER_URL;
+}
